requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
packaging==25.0
promise==2.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
Django>=5.0
djangorestframework
drf-yasg
graphene-django
requests
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
